requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-google-genai
langchain-mistralai
langgraph
openai
google-generativeai
pydantic
python-dotenv
aiohttp
httpx
pymongo
motor
supabase
python-jose[cryptography]
phonenumbers